+++ app/Http/Controllers/GalleryTagController.php
@@ -4,12 +4,17 @@ namespace App\Http\Controllers;
 
 use App\Models\Tag;
 use App\Models\Gallery;
+use Illuminate\Support\Facades\Gate;
 
 class GalleryTagController extends Controller
 {
 
     public function attachTag(Gallery $gallery, Tag $tag)
     {
+        if (Gate::denies('connect_tag_to_gallery', $gallery)) {
+            abort(403, "Action forbidden");
+        }
+
         $gallery->tags()->attach($tag->id);
 
         return back()->with([
@@ -22,6 +27,10 @@ class GalleryTagController extends Controller
 
     public function detachTag(Gallery $gallery, Tag $tag)
     {
+        if (Gate::denies('connect_tag_to_gallery', $gallery)) {
+            abort(403, "Action forbidden");
+        }
+
         $gallery->tags()->detach($tag->id);
 
         return back()->with([
